fixed direction for moving to new rooms, gel movement, sped up player…

+++ LoZGame/Enemies/EnemyClasses/EnemyEssentials.cs
@@ -21,7 +21,7 @@
 
         public Color CurrentTint { get; set; }
 
-        public int MoveSpeed { get; set; }
+        public float MoveSpeed { get; set; }
 
         public int Damage { get; set; }
